+++ packages/tests/server/interop/react/__testHelpers.tsx
@@ -14,8 +14,8 @@ import {
   splitLink,
   wsLink,
 } from '@trpc/client/src';
+import { OutputWithCursor } from '@trpc/react-query/shared';
 import { createReactQueryHooks } from '@trpc/react-query/src';
-import { OutputWithCursor } from '@trpc/react-query/src/shared/hooks/createHooksInternal';
 import * as trpcServer from '@trpc/server/src';
 import { TRPCError } from '@trpc/server/src';
 import { observable } from '@trpc/server/src/observable';
