+++ packages/tests/server/react/__testHelpers.tsx
@@ -10,7 +10,7 @@ import {
   wsLink,
 } from '@trpc/client';
 import { createTRPCReact } from '@trpc/react-query';
-import { OutputWithCursor } from '@trpc/react-query/src/shared/hooks/createHooksInternal';
+import { OutputWithCursor } from '@trpc/react-query/shared';
 import { TRPCError, initTRPC } from '@trpc/server';
 import { observable } from '@trpc/server/src/observable';
 import { subscriptionPullFactory } from '@trpc/server/src/subscription';
